Hien thi danh sach bien lai, them chuc nang thanh toan

+++ app/src/main/java/java/android/quanlybanhang/HelperClasses/Package_AdapterNhanVien/AdapterNhanVien.java
@@ -7,12 +7,14 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.squareup.picasso.Picasso;
 
+import java.android.quanlybanhang.Common.ThongTinCuaHangSql;
 import java.android.quanlybanhang.Model.NhanVien_CaLam.NhanVien;
 import java.android.quanlybanhang.R;
 import java.android.quanlybanhang.function.NhanVien.ActivityUpdateNhanVien;
@@ -27,19 +29,25 @@ public class AdapterNhanVien extends RecyclerView.Adapter<AdapterNhanVien.NhanVi
     private ArrayList<NhanVien> arrayList;
     private ListNhanVien context;
     private Context context1;
+    private ThongTinCuaHangSql thongTinCuaHangSql;
+    private boolean isChu;
+    private NhanVien nhanVien;
 
-    public AdapterNhanVien(Context context1, ListNhanVien context, ArrayList<NhanVien> arrayList){
+    public AdapterNhanVien(Context context1, ListNhanVien context, ArrayList<NhanVien> arrayList) {
         this.arrayList = arrayList;
         this.context = context;
         this.context1 = context1;
+        thongTinCuaHangSql = new ThongTinCuaHangSql(context);
+        isChu = thongTinCuaHangSql.isChu();
+        nhanVien = thongTinCuaHangSql.selectUser();
     }
 
     @NonNull
     @Override
     public NhanVienViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.item_listnhanvien,parent,false);
+        View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.item_listnhanvien, parent, false);
         NhanVienViewHolder nhanVienViewHolder = new NhanVienViewHolder(view);
-        return nhanVienViewHolder ;
+        return nhanVienViewHolder;
     }
 
     @Override
@@ -65,7 +73,7 @@ public class AdapterNhanVien extends RecyclerView.Adapter<AdapterNhanVien.NhanVi
         private ImageView imgSua;
         private ImageView imgXoa;
         private CircleImageView usernhanvien;
-        private  View view;
+        private View view;
 
 
         public NhanVienViewHolder(@NonNull View itemView) {
@@ -81,9 +89,17 @@ public class AdapterNhanVien extends RecyclerView.Adapter<AdapterNhanVien.NhanVi
                 @Override
                 public void onClick(View v) {
                     int position = getLayoutPosition();
-                    Intent intent = new Intent(context1, ActivityUpdateNhanVien.class);
-                    intent.putExtra("Key_arrayNV", arrayList.get(position));
-                    context1.startActivity(intent);
+                    if (arrayList.get(position).isChuCuaHang() && isChu) {
+                        Intent intent = new Intent(context1, ActivityUpdateNhanVien.class);
+                        intent.putExtra("Key_arrayNV", arrayList.get(position));
+                        context1.startActivity(intent);
+                    } else if (!arrayList.get(position).isChuCuaHang() && nhanVien.getChucVu().get(0) && !nhanVien.getEmail().equals(arrayList.get(position).getEmail())) {
+                        Intent intent = new Intent(context1, ActivityUpdateNhanVien.class);
+                        intent.putExtra("Key_arrayNV", arrayList.get(position));
+                        context1.startActivity(intent);
+                    } else {
+                        Toast.makeText(context1, "Không thực hiện được hành động này", Toast.LENGTH_SHORT).show();
+                    }
                 }
             });
 
@@ -91,9 +107,19 @@ public class AdapterNhanVien extends RecyclerView.Adapter<AdapterNhanVien.NhanVi
                 @Override
                 public void onClick(View v) {
                     int position = getLayoutPosition();
-                    Intent intent = new Intent(context1, ChamCongNhanVienActivity.class);
-                    intent.putExtra("ID_NHANVIEN", arrayList.get(position).getId());
-                    context1.startActivity(intent);
+                    if (arrayList.get(position).isChuCuaHang() && isChu) {
+                        Intent intent = new Intent(context1, ChamCongNhanVienActivity.class);
+                        intent.putExtra("ID_NHANVIEN", arrayList.get(position).getId());
+                        context1.startActivity(intent);
+                    } else if (!arrayList.get(position).isChuCuaHang() && nhanVien.getChucVu().get(0) && !nhanVien.getEmail().equals(arrayList.get(position).getEmail())) {
+                        Intent intent = new Intent(context1, ChamCongNhanVienActivity.class);
+                        intent.putExtra("ID_NHANVIEN", arrayList.get(position).getId());
+                        context1.startActivity(intent);
+                    }
+                    else {
+                        Toast.makeText(context1, "Không thực hiện được hành động này", Toast.LENGTH_SHORT).show();
+                    }
+
                 }
             });
 
@@ -101,7 +127,13 @@ public class AdapterNhanVien extends RecyclerView.Adapter<AdapterNhanVien.NhanVi
                 @Override
                 public void onClick(View v) {
                     int position = getLayoutPosition();
-                    context.delete(position);
+                    if (arrayList.get(position).isChuCuaHang() && isChu) {
+//                        context.delete(position);
+                    } else if (!arrayList.get(position).isChuCuaHang() && nhanVien.getChucVu().get(0) && !nhanVien.getEmail().equals(arrayList.get(position).getEmail())) {
+                        context.delete(position);
+                    } else {
+                        Toast.makeText(context1, "Không thực hiện được hành động này", Toast.LENGTH_SHORT).show();
+                    }
                 }
             });
         }
